+++ src/plugins/index.ts
@@ -1,6 +1,6 @@
 export { intersect } from './intersect';
 export { mask } from './mask';
 export { transition } from './transition';
-export { persist } from './persist';
+export { persist, persistStore } from './persist';
 export { focus } from './focus';
 export { collapse } from './collapse';